Make code snippet a bit easier to copy-paste
Helps to avoid that people get stuck in the tutorial.

At the moment it seems like the ng-repeat section should be deleted. So the comment `<!-- ... --> should be one line above the closing tag, but it is much easier to just paste all the code.

This way it is easier to copy-paste without the need to think about which exact line has to be put where.